+++ example/src/views/Komponenter/RadioknapExample.vue
@@ -2,113 +2,55 @@
   <section>
     <fds-component-preview header="Eksempel">
       <fds-formgroup>
-        <fds-label>Vælg radio</fds-label>
+        <fds-label>Vælg Genre</fds-label>
 
-        <fds-radio header="Pick one" :list="radioOptions" v-model="radioVal">
-          <template v-slot:[`melon`]>
-            <p>Det er muligt at benytte radio til mere indhold</p>
-          </template>
-        </fds-radio>
+        <fds-radio-group v-model="radioValueKerne">
+          <fds-radio-item value="metal"> Metal </fds-radio-item>
+          <fds-radio-item value="pop">
+            Pop
+            <template #content>
+              Pop er som musikgenre den mest udbredte musikstilart i moderne tid - væsentlig mere
+              udbredt end fx rock, jazz og folkemusik. Navnet er en afledning af ordet "populær"
+              (eng. "popular").
+            </template>
+          </fds-radio-item>
+          <fds-radio-item value="klassisk"> Klassist </fds-radio-item>
+        </fds-radio-group>
       </fds-formgroup>
 
-      <fds-pre header="v-model" :json="{ radioVal }" />
+      <fds-pre :json="{ radioValueKerne }" />
 
       <template #description>
-        <p class="italic">
-          <code>fds-radio</code> Ud fra en liste af <code>FdsOptionItem[]</code> genereres radio
-          knapperne.
-        </p>
-        <p class="italic">
-          Det er muligt at udfolde valgte radio og angive eget indhold, via dynamisk
-          <code>slot</code>
-        </p>
+        <p class="italic">Det er muligt at udfolde valgte radio og angive eget indhold</p>
       </template>
       <template #code>
         <pre v-text="code"></pre>
-      </template>
-    </fds-component-preview>
-
-    <hr class="my-6" />
-
-    <fds-component-preview header="Ja eller nej">
-      <fds-formgroup>
-        <fds-label>Vælg radio toggle</fds-label>
-
-        <fds-radio-toggle v-model="toggleRadio">
-          <template v-slot:[`true`]>
-            <p>Det er muligt at benytte radio til mere indhold</p>
-          </template>
-        </fds-radio-toggle>
-      </fds-formgroup>
-
-      <fds-pre header="v-model" :json="{ toggleRadio }" />
-
-      <template #description>
-        <p class="italic">
-          Er det kun ja eller nej - sand eller falsk, kan <code>fds-radio-toggle</code> benyttes.
-        </p>
-        <p class="italic">
-          Det er muligt at udfolde valgte radio og angive eget indhold, via dynamisk
-          <code>slot</code>
-        </p>
-      </template>
-      <template #code>
-        <pre v-text="code2"></pre>
       </template>
     </fds-component-preview>
   </section>
 </template>
 
 <script setup lang="ts">
-import { FdsOptionItem } from 'dkfds-vue3/src/model/fds.model';
 import { ref } from 'vue';
 
-const radioVal = ref('');
-
-const toggleRadio = ref(null);
-const radioOptions = ref<FdsOptionItem[]>([
-  {
-    title: 'Banan',
-    value: 'banan',
-  },
-  {
-    title: 'Melon',
-    value: 'melon',
-  },
-  {
-    title: 'Æble',
-    value: 'æble',
-  },
-]);
+const radioValueKerne = ref('');
 
 const code = `
-<fds-radio header="Pick one" :list="radioOptions" v-model="radioVal">
-  <template v-slot:[\`melon\`]>
-    <p>Det er muligt at benytte radio til mere indhold</p>
-  </template>
-</fds-radio>
+<fds-formgroup>
+  <fds-label>Vælg Genre</fds-label>
 
-const radioOptions = ref<FdsOptionItem[]>([
-  {
-    title: 'Banan',
-    value: 'banan',
-  },
-  {
-    title: 'Melon',
-    value: 'melon',
-  },
-  {
-    title: 'Æble',
-    value: 'æble',
-  },
-]);
-`;
-
-const code2 = `
-<fds-radio-toggle v-model="toggleRadio">
-  <template v-slot:[\`true\`]>
-    <p>Det er muligt at benytte radio til mere indhold</p>
-  </template>
-</fds-radio-toggle>
+  <fds-radio-group v-model="radioValueKerne">
+    <fds-radio-item value="metal"> Metal </fds-radio-item>
+    <fds-radio-item value="pop">
+      Pop
+      <template #content>
+        Pop er som musikgenre den mest udbredte musikstilart i moderne tid - væsentlig mere
+        udbredt end fx rock, jazz og folkemusik. Navnet er en afledning af ordet "populær"
+        (eng. "popular").
+      </template>
+    </fds-radio-item>
+    <fds-radio-item value="klassisk"> Klassist </fds-radio-item>
+  </fds-radio-group>
+</fds-formgroup>
 `;
 </script>